pyproject.toml
[tool.poetry]
name = "arabic-vocab-tester"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.0.5"
arabic-reshaper = "^2.1.0"
python-bidi = "^0.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
